Introduce the capability to log stack traces only at certain levels.
The rationale for wanting this feature is that usually the readers of the
log are not exactly interested in the stack traces unless they are
developers (which, most of the time, they are not).

Stacktraces fill up the log and occlude the possibly important messages
in "cruft".

Therefore it would be nice to NOT log stack traces by default but enable
the logging system to do so on demand.

The imagined workflow goes like this:

1) User encounters a strange message in the logs.
2) User calls up support.
3) Support doesn't know how this should be solved, calls up the developer.
4) Developer is not sure how this could happen and would like to see the
stacktrace.
5) Support asks the user to switch on the debug logging on a certain
logger and replicate the issue.
6) All of the sudden the stack traces appear along with the very same
messages without the need to change the code.
7) Developer sees what he wanted, fixes the issue.